+++ examples/sptrsm_helpers.hpp
@@ -44,7 +44,7 @@
     long flops_per_iter =                                                      \
         n_vectors * (crs_mat_D_plus_L->nnz * SPTRSM_FLOPS_PER_NZ +             \
                      crs_mat_D_plus_L->n_rows * SPTRSM_FLOPS_PER_ROW);         \
-    long iter_per_second = static_cast<long>(n_iter / runtime);                \
+    double iter_per_second = n_iter / runtime;                                 \
                                                                                \
     std::cout << "Performance: " << flops_per_iter * iter_per_second * F_TO_GF \
               << " [GF/s]" << std::endl;                                       \
